renames and file consolidation
- rename `interface` -> `api`
- move contents of `interface` into `api`
- rename the individual selectors to prefix them with `sel_`

This just makes it a bit easier to see at a glance which files do what:

- `mod.rs` is just the mod declarations
- `api.rs` is everything public to other modules
- `base.rs` is the base classes for selectors
- `sel_<foo>` is a selector for `<foo>`
