+++ src/select/mod.rs
@@ -1,6 +1,5 @@
+mod api;
 mod base;
-mod interface;
-mod list_item;
-mod section;
-pub mod selector;
-pub use interface::*;
+mod sel_list_item;
+mod sel_section;
+pub use api::*;
